feat : refactor admin panel and add view HCP and approve card components

+++ components/base/admin-dashboard/data-table/healthcare-providers-table/actions.tsx
@@ -11,6 +11,17 @@ import { User } from "@prisma/client";
 import { CheckIcon, EyeOpenIcon, TrashIcon } from "@radix-ui/react-icons";
 import { toast } from "sonner";
 
+import {
+  AlertDialog,
+  AlertDialogAction,
+  AlertDialogCancel,
+  AlertDialogContent,
+  AlertDialogDescription,
+  AlertDialogFooter,
+  AlertDialogHeader,
+  AlertDialogTitle,
+  AlertDialogTrigger,
+} from "@/components/ui/alert-dialog";
 import { Button } from "@/components/ui/button";
 import { Spinner } from "@/components/ui/spinner";
 
@@ -31,10 +42,17 @@ export function HealthcareProvidersDataTableActions({
 }: HealthcareProvidersDataTableActionsProps) {
   const router = useRouter();
   const [isPending, startTransition] = React.useTransition();
-
+  const healthcareProviderEmail = healthcareProvider!.user.email;
+  const healthcareProviderName = healthcareProvider!.user.name;
   const handleVerify = (id: string) => {
     startTransition(() => {
-      verifyHealthcareProvider(id).then((data) => {
+      verifyHealthcareProvider(
+        id,
+        healthcareProviderEmail,
+        "oladoc-customer-service",
+        "Your oladoc pro account is now verified!",
+        healthcareProviderName,
+      ).then(async (data) => {
         if (data?.error) {
           toast.error(data.error);
           return;
@@ -48,7 +66,13 @@ export function HealthcareProvidersDataTableActions({
 
   const handleDelete = (id: string) => {
     startTransition(() => {
-      deleteHealthcareProvider(id).then((data) => {
+      deleteHealthcareProvider(
+        id,
+        healthcareProviderEmail,
+        "oladoc-customer-service",
+        "Your oladoc pro account has been deleted!",
+        healthcareProviderName,
+      ).then((data) => {
         if (data?.error) {
           toast.error(data.error);
           return;
@@ -83,16 +107,38 @@ export function HealthcareProvidersDataTableActions({
         {isPending ? <Spinner /> : <CheckIcon className="h-4 w-4" />}
         <span>Verify</span>
       </Button>
-      <Button
-        size="sm"
-        variant="destructive"
-        className="gap-1"
-        onClick={() => handleDelete(healthcareProvider.id)}
-        disabled={isPending}
-      >
-        {isPending ? <Spinner /> : <TrashIcon className="h-4 w-4" />}
-        <span>Delete</span>
-      </Button>
+
+      <AlertDialog>
+        <AlertDialogTrigger asChild>
+          <Button
+            size="sm"
+            variant="destructive"
+            className="gap-1"
+            disabled={isPending}
+          >
+            {isPending ? <Spinner /> : <TrashIcon className="h-4 w-4" />}Delete
+          </Button>
+        </AlertDialogTrigger>
+        <AlertDialogContent>
+          <AlertDialogHeader>
+            <AlertDialogTitle>
+              Are you sure you want to delete this account?
+            </AlertDialogTitle>
+            <AlertDialogDescription>
+              This action cannot be undone. This will permanently delete this
+              account and remove the data from our servers.
+            </AlertDialogDescription>
+          </AlertDialogHeader>
+          <AlertDialogFooter>
+            <AlertDialogCancel className="mt-0">Cancel</AlertDialogCancel>
+            <AlertDialogAction
+              onClick={() => handleDelete(healthcareProvider.id)}
+            >
+              Continue
+            </AlertDialogAction>
+          </AlertDialogFooter>
+        </AlertDialogContent>
+      </AlertDialog>
     </div>
   );
 }
